# [Fix] Fix potential circular import in core, data and utils, disconnect unnecessary pytorch3d dependency