--- a/demo/estimate_smpl.py
+++ b/demo/estimate_smpl.py
@@ -15,7 +15,7 @@ from mmhuman3d.apis import (
     inference_video_based_model,
     init_model,
 )
-from mmhuman3d.core.visualization import visualize_smpl_hmr
+from mmhuman3d.core.visualization.visualize_smpl import visualize_smpl_hmr
 from mmhuman3d.data.data_structures.human_data import HumanData
 from mmhuman3d.utils import array_to_images
 from mmhuman3d.utils.demo_utils import (