--- a/mmhuman3d/__init__.py
+++ b/mmhuman3d/__init__.py
@@ -16,7 +16,7 @@ def digit_version(version_str):
 
 
 mmcv_minimum_version = '1.3.1'
-mmcv_maximum_version = '1.5.1'
+mmcv_maximum_version = '1.5.2'
 mmcv_version = digit_version(mmcv.__version__)