--- a/mmhuman3d/apis/train.py
+++ b/mmhuman3d/apis/train.py
@@ -15,7 +15,7 @@ from mmhuman3d.core.distributed_wrapper import DistributedDataParallelWrapper
 from mmhuman3d.core.evaluation import DistEvalHook, EvalHook
 from mmhuman3d.core.optimizer import build_optimizers
 from mmhuman3d.data.datasets import build_dataloader, build_dataset
-from mmhuman3d.utils import get_root_logger
+from mmhuman3d.utils.logger import get_root_logger
 
 
 def set_random_seed(seed, deterministic=False):