--- a/mmhuman3d/core/renderer/torch3d_renderer/mesh_renderer.py
+++ b/mmhuman3d/core/renderer/torch3d_renderer/mesh_renderer.py
@@ -1,29 +1,26 @@
-from typing import Iterable, Optional, Union
+from typing import Iterable, Optional, Tuple, Union
 
 import torch
 from pytorch3d.structures import Meshes
 
 from mmhuman3d.core.cameras import MMCamerasBase
 from .base_renderer import BaseRenderer
-from .builder import RENDERER
-from .utils import normalize
+from .lights import MMLights
 
 
-@RENDERER.register_module(
-    name=['Normal', 'normal', 'normal_renderer', 'NormalRenderer'])
-class NormalRenderer(BaseRenderer):
-    """Render normal map with the help of camera system."""
-    shader_type = 'NormalShader'
+class MeshRenderer(BaseRenderer):
+    """Render RGBA image with the help of camera system."""
+    shader_type = 'SoftPhongShader'
 
     def __init__(
         self,
-        resolution: Iterable[int] = None,
+        resolution: Tuple[int, int] = None,
         device: Union[torch.device, str] = 'cpu',
         output_path: Optional[str] = None,
         out_img_format: str = '%06d.png',
         **kwargs,
     ) -> None:
-        """Renderer for normal map of meshes.
+        """Renderer for RGBA image of meshes.
 
         Args:
             resolution (Iterable[int]):
@@ -37,33 +34,30 @@ class NormalRenderer(BaseRenderer):
             out_img_format (str, optional): The image format string for
                 saving the images.
                 Defaults to '%06d.png'.
-
-        Returns:
-            None
         """
         super().__init__(
             resolution=resolution,
             device=device,
             output_path=output_path,
-            obj_path=None,
             out_img_format=out_img_format,
             **kwargs)
 
     def forward(self,
-                meshes: Optional[Meshes] = None,
+                meshes: Meshes,
                 cameras: Optional[MMCamerasBase] = None,
+                lights: Optional[MMLights] = None,
                 indexes: Optional[Iterable[int]] = None,
                 backgrounds: Optional[torch.Tensor] = None,
-                **kwargs):
+                **kwargs) -> Union[torch.Tensor, None]:
         """Render Meshes.
 
         Args:
-            meshes (Optional[Meshes], optional): meshes to be rendered.
-                Defaults to None.
+            meshes (Meshes): meshes to be rendered.
             cameras (Optional[MMCamerasBase], optional): cameras for render.
                 Defaults to None.
-            indexes (Optional[Iterable[int]], optional): indexes for the
-                images.
+            lights (Optional[MMLights], optional): lights for render.
+                Defaults to None.
+            indexes (Optional[Iterable[int]], optional): indexes for images.
                 Defaults to None.
             backgrounds (Optional[torch.Tensor], optional): background images.
                 Defaults to None.
@@ -75,17 +69,14 @@ class NormalRenderer(BaseRenderer):
         meshes = meshes.to(self.device)
         self._update_resolution(cameras, **kwargs)
         fragments = self.rasterizer(meshes_world=meshes, cameras=cameras)
-        normal_map = self.shader(
-            fragments=fragments, meshes=meshes, cameras=cameras)
+
+        rendered_images = self.shader(
+            fragments=fragments,
+            meshes=meshes,
+            cameras=cameras,
+            lights=self.lights if lights is None else lights)
 
         if self.output_path is not None:
-            rgba = self.tensor2rgba(normal_map)
+            rgba = self.tensor2rgba(rendered_images)
             self._write_images(rgba, backgrounds, indexes)
-
-        return normal_map
-
-    def tensor2rgba(self, tensor: torch.Tensor):
-        rgbs, valid_masks = tensor[..., :3], (tensor[..., 3:] > 0) * 1.0
-        rgbs = normalize(
-            rgbs, origin_value_range=(-1, 1), out_value_range=(0, 1))
-        return torch.cat([rgbs, valid_masks], -1)
+        return rendered_images