--- a/mmhuman3d/core/renderer/torch3d_renderer/meshes.py
+++ b/mmhuman3d/core/renderer/torch3d_renderer/meshes.py
@@ -11,7 +11,7 @@ from mmhuman3d.models.body_models.builder import SMPL, SMPLX
 from mmhuman3d.utils.mesh_utils import \
     join_meshes_as_batch as _join_meshes_as_batch
 from .builder import build_renderer
-from .textures import TexturesNearest
+from .textures.textures import TexturesNearest
 from .utils import align_input_to_padded