--- a/mmhuman3d/core/renderer/torch3d_renderer/pointcloud_renderer.py
+++ b/mmhuman3d/core/renderer/torch3d_renderer/pointcloud_renderer.py
@@ -13,13 +13,8 @@ from pytorch3d.structures import Meshes, Pointclouds
 from mmhuman3d.core.cameras import MMCamerasBase
 from mmhuman3d.utils.mesh_utils import mesh_to_pointcloud_vc
 from .base_renderer import BaseRenderer
-from .builder import RENDERER
 
 
-@RENDERER.register_module(name=[
-    'PointCloud', 'pointcloud', 'point_cloud', 'pointcloud_renderer',
-    'PointCloudRenderer'
-])
 class PointCloudRenderer(BaseRenderer):
 
     def __init__(self,