--- a/mmhuman3d/core/renderer/torch3d_renderer/shader/__init__.py
+++ b/mmhuman3d/core/renderer/torch3d_renderer/shader/__init__.py
@@ -1,0 +1,16 @@
+# yapf: disable
+from .builder import (  # noqa: F401
+    DepthShader,
+    HardFlatShader,
+    HardGouraudShader,
+    HardPhongShader,
+    NoLightShader,
+    NormalShader,
+    SegmentationShader,
+    SilhouetteShader,
+    SoftGouraudShader,
+    SoftPhongShader,
+    build_shader,
+)
+
+# yapf: enable