--- a/mmhuman3d/core/renderer/torch3d_renderer/uv_renderer.py
+++ b/mmhuman3d/core/renderer/torch3d_renderer/uv_renderer.py
@@ -20,11 +20,9 @@ from mmhuman3d.core.cameras.cameras import (
     MMCamerasBase,
 )
 from mmhuman3d.utils.path_utils import check_path_suffix
-from .builder import RENDERER
 from .utils import array2tensor, rgb2bgr
 
 
-@RENDERER.register_module(name=['uv_renderer', 'uv', 'UV', 'UVRenderer'])
 class UVRenderer(nn.Module):
     """Renderer for SMPL(x) UV map."""