@LazyBusyYang 2022-05-30
Import time compare:
<!DOCTYPE html>

import content | import time(main, sec) | import time(PR#190, sec) | speed-up (x)
-- | -- | -- | --
convert_kps | 1.9779 | 1.9305 | 1.0246
visualize_kp2d | 5.0012 | 3.7096 | 1.3482
visualize_kp3d | 4.8724 | 3.7236 | 1.3085
images_to_array | 2.7962 | 2.0348 | 1.3742
check_path_suffix | 2.8509 | 1.9142 | 1.4893
HumanData | 5.0130 | 1.8973 | 2.6422
visualize_smpl_calibration | 4.8914 | 4.0477 | 1.2084
MeshRenderer | 5.0428 | 1.8929 | 2.6641
ParametricMeshes | 5.0820 | 2.7389 | 1.8555
MMLights | 4.8773 | 1.9466 | 2.5055
build_shader | 4.7687 | 2.0013 | 2.3828
build_renderer | 5.0584 | 2.8075 | 1.8017