@LazyBusyYang 2022-05-30
Import failure compare if pytorch3d is not installed

S: success, F: failure, W: success with a warning.

<!DOCTYPE html>

import content | result (main, S/F/W) | result (PR#190, S/F/W)
-- | -- | --
convert_kps | S | S
visualize_kp2d | F | W
visualize_kp3d | F | W
images_to_array | F | S
check_path_suffix | F | S
HumanData | F | S
visualize_smpl_calibration | F | F
MeshRenderer | F | F
ParametricMeshes | F | F
MMLights | F | F
build_shader | F | F
build_renderer | F | F