@caizhongang 2022-05-31
Is this [file](https://github.com/open-mmlab/mmhuman3d/blob/main/tools/test_model_builders.py) meant to be included in the main branch?