@LazyBusyYang 2022-05-31
> Is this [file](https://github.com/open-mmlab/mmhuman3d/blob/main/tools/test_model_builders.py) meant to be included in the main branch?



> d7079a3

No. Thanks for finding this mistake! Removed in [d7079a3](https://github.com/open-mmlab/mmhuman3d/pull/190/commits/d7079a3855776475ebc775c7f671a415b2318b36).